Change file structure and split classes in different files
